Modify cat/catdto weigth property to double. Finish the dynamic database.

+++ Actividad2/Actividad2/Domain/Dto/CatDto.cs
@@ -5,7 +5,7 @@ namespace Actividad2.Domain.Dto;
 public class CatDto
 {
     public CatDto(){}
-    public CatDto(Guid id, string name, int age, string race, int weight, HealthState healthState, Guid colonyId)
+    public CatDto(Guid id, string name, int age, string race, double weight, HealthState healthState, Guid colonyId)
     {
         Id = id;
         Name = name;
@@ -20,7 +20,7 @@ public class CatDto
     public string Name { get; set; }
     public int Age { get; set; }
     public string Race  { get; set; }
-    public int Weight { get; set; }
+    public double Weight { get; set; }
     public HealthState HealthState { get; set; }
     public Guid ColonyId { get; set; }
 }
